made pacman.h and modified scene.h, scene.cpp and pacman.cpp

+++ practical_4/scene.h
@@ -8,7 +8,7 @@ public:
 	Scene() = default;
 
 	virtual ~Scene() = default;
-	virtual void update(double dt);
+	virtual void update(float dt);
 	virtual void render();
 	virtual void load() = 0;
 	std::vector<std::shared_ptr<Entity>> &getEnts();
